requirements.txt
aniso8601==9.0.1
APScheduler==3.8.1
backports.zoneinfo==0.2.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
DateTime==4.3
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-APScheduler==1.12.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-RESTful==0.3.9
Flask-WTF==0.15.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
lxml==4.6.3
MarkupSafe==2.0.1
mongoengine==0.23.1
multitasking==0.0.9
numpy==1.21.2
pandas==1.3.3
pycparser==2.20
PyJWT==2.3.0
pymongo==3.12.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
tzdata==2021.5
tzlocal==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
WTForms==2.3.3
yfinance==0.1.63
zope.interface==5.4.0
